+++ R/get_parsed_data.R
@@ -19,12 +19,18 @@ get_parsed_data <- function(file = "",
   } else {
     script <- text
   }
+  
+  if (is.null(script)) {
+    script <- readLines(file)
+  }
+  # script <- gsub("\\", "\\\\", script, fixed = TRUE)
+  
   # detailed information about parsed code
   dat <- utils::getParseData(parse(file = file,
                                    text = script,
                                    keep.source = TRUE),
                              includeText = TRUE)
-  dat <- dat[dat$terminal == TRUE,
+  dat <- dat[dat$terminal == TRUE, # Exclude Linting - more verbose
              c("line1", "col1", "line2", "col2", "token", "text")]
 
   # replace wiht original line indices