+++ tests/testthat/test-check_pkg_usage.R
@@ -43,12 +43,12 @@ testthat::test_that("check_pkg_usage() working on test", {
                            ":=", "as.IDate", "as.data.table",
                            "copy", "is.data.table", "last",
                            "setkey", "setkey", "%>%", "bind_cols",
-                           "filter", "id", "if_else", "mutate",
+                           "filter", "if_else", "mutate",
                            "n", "n_distinct", "map",
                            "bind_cols_2", NA))
   testthat::expect_equal(res$n_calls,
                          c(2, 10, 1, 2, 2, 1, 1, 1, 1, 1, 4, 3,
-                           3, 7, 10, 7, 1, 1, 3, 6,
+                           3, 7, 10, 8, 1, 3, 6,
                            5, 4, 1, 0))
 
   # run function with markers
@@ -113,7 +113,7 @@ testthat::test_that("check_pkg_usage working on target", {
                            ":=", "as.IDate", "as.data.table",
                            "copy", "is.data.table", "last",
                            "setkey", "bind_cols", "filter",
-                           "id", "if_else", "mutate", "n",
+                           "if_else", "mutate", "n",
                            "n_distinct", "map", "%>%", "bind_cols_2"
                          ))
 
@@ -172,15 +172,14 @@ testthat::test_that("check_pkg_usage fails - empty files", {
   writeLines(character(3), con = test_file_empty)
 
   testthat::expect_message(
-    testthat::expect_equal(
+    testthat::expect_null(
       check_pkg_usage(path = dir,
                       pkgs = c("data.table",
                                "dplyr",
                                "purrr"
                       ),
                       recursive = FALSE,
-                      use_markers = FALSE),
-      NULL),
+                      use_markers = FALSE)),
     "All provided files are empty")
 })